backend/requirements.txt
fastapi
uvicorn
langchain-google-genai==1.0.7
langchain-community==0.2.6
chromadb==0.5.3
python-dotenv==1.0.1
pymupdf==1.24.5
python-docx==1.1.2
google-genai
sentence-transformers
pydantic==2.10.1